Avoid creating unnecessary thread pool
Unless the `threadpool` feature is disabled, `Server` always contains a
thread pool, even if it is unnecessary because the `Server` is only used
with `serve_single_thread`. Fix this by deferring creating the thread
pool until it is actually required.

I noticed this because I was getting resource exhaustion errors when
running a single-threaded server in a podman container, which restricts
the number of threads. Disabling the `threadpool` feature also fixes
this, but this must be done explicitly since it's a default feature.